+++ SpaghettoCosmosTestKernel.Plugs/SpaghettoCosmosTestKernel.Plugs.csproj
@@ -1,0 +1,29 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <TargetFramework>net6.0</TargetFramework>
+    <ImplicitUsings>enable</ImplicitUsings>
+    <Nullable>enable</Nullable>
+  </PropertyGroup>
+
+	<ItemGroup>
+		<PackageReference Include="Cosmos.System2" Version="0-*" NoWarn="NU1604"></PackageReference>
+	</ItemGroup>
+
+	<ItemGroup>
+	  <ProjectReference Include="..\SpaghettoCosmosTestKernel\SpaghettoCosmosTestKernel.csproj" />
+	</ItemGroup>
+
+	<ItemGroup>
+	  <Reference Include="IL2CPU">
+	    <HintPath>..\..\..\..\AppData\Roaming\Cosmos User Kit\Build\IL2CPU\IL2CPU.dll</HintPath>
+	  </Reference>
+	  <Reference Include="IL2CPU.API">
+	    <HintPath>..\..\..\..\AppData\Roaming\Cosmos User Kit\Build\IL2CPU\IL2CPU.API.dll</HintPath>
+	  </Reference>
+	  <Reference Include="XSharp">
+	    <HintPath>..\..\..\..\AppData\Roaming\Cosmos User Kit\Build\IL2CPU\XSharp.dll</HintPath>
+	  </Reference>
+	</ItemGroup>
+
+</Project>
