+++ spaghetto/BuiltinTypes/SBuiltinType.cs
@@ -1,0 +1,13 @@
+﻿namespace spaghetto {
+    public enum SBuiltinType
+    {
+        String,
+        Int,
+        Float,
+        List,
+        Null,
+        NativeFunc,
+        Function,
+        NativeLibraryImporter,
+    }
+}
