+++ spaghetto/Parsing/Parser.cs
@@ -1173,472 +1173,4 @@ namespace spaghetto {
             return "TokenNode: " + Token.Type.ToString() + " val=" + Token.Value?.ToString() + " text=" + Token.Text.ToString();
         }
     }
-
-    public abstract class SValue
-    {
-        public static SValue Null => new SNull();
-        public abstract SBuiltinType BuiltinName { get; }
-
-        #region Metadata
-        internal bool TypeIsFixed { get; set; } = true;
-        #endregion
-
-        public virtual SValue Add(SValue other)
-        {
-            throw NotSupportedOn("Add");
-        }
-
-        public virtual SValue Sub(SValue other)
-        {
-            throw NotSupportedOn("Sub");
-        }
-
-        public virtual SValue Mul(SValue other)
-        {
-            throw NotSupportedOn("Mul");
-        }
-
-        public virtual SValue Div(SValue other)
-        {
-            throw NotSupportedOn("Div");
-        }
-
-        public virtual SValue Mod(SValue other)
-        {
-            throw NotSupportedOn("Mod");
-        }
-
-        public virtual SValue Idx(SValue other)
-        {
-            throw NotSupportedOn("Idx");
-        }
-
-        public virtual SValue Dot(SValue other) {
-            throw NotSupportedOn("Dot");
-        }
-
-        public virtual SValue Equals(SValue other) {
-            throw NotSupportedOn("Equals");
-        }
-
-        public virtual SValue LessThan(SValue other) {
-            throw NotSupportedOn("LessThan");
-        }
-
-        public virtual SValue LessThanEqu(SValue other) {
-            throw NotSupportedOn("LessThanEqu");
-        }
-
-        public virtual SValue GreaterThan(SValue other) {
-            throw NotSupportedOn("GreaterThan");
-        }
-
-        public virtual SValue GreaterThanEqu(SValue other) {
-            throw NotSupportedOn("GreaterThanEqu");
-        }
-
-        public virtual SValue CastToBuiltin(SBuiltinType other) {
-            throw NotSupportedOn("CastToBuiltin");
-        }
-
-        public virtual SValue Call(Scope scope, List<SValue> args) {
-            throw NotSupportedOn("Call");
-        }
-
-        public virtual bool IsNull() {
-            return false;
-        }
-
-        public virtual SValue Not()
-        {
-            throw NotSupportedOn("Not");
-        }
-
-        public virtual SValue ArithNot()
-        {
-            throw NotSupportedOn("ArithNot");
-        }
-
-        public abstract bool IsTruthy();
-
-        public override string ToString()
-        {
-            return BuiltinName.ToString();
-        }
-
-        public virtual SString ToSpagString()
-        {
-            return new SString("<unknown of type " + BuiltinName.ToString() + ">");
-        }
-
-        protected NotImplementedException NotSupportedBetween(SValue other, string type) {
-            return new NotImplementedException(type + " not supported between " + GetType().Name + " and " + other.GetType().Name);
-        }
-
-        protected NotImplementedException NotSupportedOn(string type) {
-            return new NotImplementedException(type + " is not supported on " + GetType().Name);
-        }
-
-        protected ArgumentException CastInvalid(string type) {
-            return new ArgumentException(GetType().Name + " can not be cast to " + type);
-        }
-
-        internal void CopyMeta(ref SValue other) {
-            other.TypeIsFixed = TypeIsFixed;
-        }
-    }
-
-    public class SNativeLibraryImporter : SValue {
-        public override SBuiltinType BuiltinName => SBuiltinType.NativeLibraryImporter;
-        public Action<Scope> Import { get; set; } = (Scope scope) => { };
-
-        public SNativeLibraryImporter(Action<Scope> import) {
-            Import = import;
-        }
-
-        public override bool IsTruthy() {
-            return true;
-        }
-    }
-
-    public class SNativeFunction : SValue {
-        public override SBuiltinType BuiltinName => SBuiltinType.NativeFunc;
-        public Func<Scope, List<SValue>, SValue> Impl { get; set; }
-
-        public SNativeFunction(Func<Scope, List<SValue>, SValue> impl) {
-            Impl = impl;
-        }
-
-        /// <summary>
-        /// NOTE: The scope in SNativeFunction is the calling scope, but not in SFunction!
-        /// </summary>
-        /// <param name="scope"></param>
-        /// <param name="args"></param>
-        /// <returns></returns>
-        public override SValue Call(Scope scope, List<SValue> args) {
-            return Impl(scope, args);
-        }
-
-        public override bool IsTruthy() {
-            return true;
-        }
-    }
-
-    public class SFunction : SValue {
-        public override SBuiltinType BuiltinName => SBuiltinType.Function;
-        public string FunctionName { get; set; }
-        public List<string> Args { get; set; }
-        public SyntaxNode Callback { get; set; }
-        public Scope DefiningScope { get; set; }
-
-        public SFunction(Scope definingScope, string functionName, List<string> args, SyntaxNode callback) {
-            DefiningScope = definingScope;
-            FunctionName = functionName;
-            Args = args;
-            Callback = callback;
-        }
-
-        public override SValue Call(Scope scope, List<SValue> args) {
-            if (args.Count != Args.Count) throw new Exception(FunctionName + " expected " + Args.Count + " arguments. (" + string.Join(", ", Args) + ")");
-
-            Scope funcScope = new(DefiningScope);
-            
-            for(int i = 0; i < Args.Count; i++) {
-                funcScope.Set(Args[i], args[i]);
-            }
-
-            Callback.Evaluate(funcScope);
-
-            return funcScope.ReturnValue;
-        }
-
-        public override bool IsTruthy() {
-            return true;
-        }
-    }
-
-    public class SNull : SValue {
-        public override SBuiltinType BuiltinName => SBuiltinType.Null;
-
-        public override bool IsNull() {
-            return true;
-        }
-
-        public override bool IsTruthy() {
-            return false;
-        }
-
-        public override SString ToSpagString() {
-            return new("null");
-        }
-
-        public override string ToString() {
-            return "<SNull>";
-        }
-    }
-
-    public class SInt : SValue
-    {
-        public override SBuiltinType BuiltinName => SBuiltinType.Int;
-        public int Value { get; set; }
-
-        public SInt() {
-            Value = 0;
-        }
-
-        public SInt(int value)
-        {
-            Value = value;
-        }
-
-        public override SValue Add(SValue other) {
-            if (other is not SInt otherInt) throw new Exception("Can not perform Add on SInt and " + other.GetType().Name);
-
-            return new SInt(Value + otherInt.Value);
-        }
-
-        public override SValue Sub(SValue other) {
-            if (other is not SInt otherInt) throw new Exception("Can not perform Sub on SInt and " + other.GetType().Name);
-            return new SInt(Value - otherInt.Value);
-        }
-
-        public override SValue Mul(SValue other) {
-            if (other is not SInt otherInt) throw new Exception("Can not perform Mul on SInt and " + other.GetType().Name);
-            return new SInt(Value * otherInt.Value);
-        }
-
-        public override SValue Div(SValue other) {
-            if (other is not SInt otherInt) throw new Exception("Can not perform Div on SInt and " + other.GetType().Name);
-            return new SInt(Value / otherInt.Value);
-        }
-
-        public override SValue Mod(SValue other) {
-            if (other is not SInt otherInt) throw new Exception("Can not perform Mod on SInt and " + other.GetType().Name);
-            return new SInt(Value % otherInt.Value);
-        }
-
-        public override SValue Equals(SValue other) {
-            if (other is not SInt otherInt) throw new Exception("Can not perform EqualsCheck on SInt and " + other.GetType().Name);
-            return new SInt(Value == otherInt.Value ? 1 : 0);
-        }
-
-        public override SValue LessThan(SValue other) {
-            if (other is not SInt otherInt) throw new Exception("Can not perform LessThanCheck on SInt and " + other.GetType().Name);
-            return new SInt(Value < otherInt.Value ? 1 : 0);
-        }
-
-        public override SValue LessThanEqu(SValue other) {
-            if (other is not SInt otherInt) throw new Exception("Can not perform LessThanEquCheck on SInt and " + other.GetType().Name);
-            return new SInt(Value <= otherInt.Value ? 1 : 0);
-        }
-
-        public override SValue GreaterThan(SValue other) {
-            if (other is not SInt otherInt) throw new Exception("Can not perform GreaterThanCheck on SInt and " + other.GetType().Name);
-            return new SInt(Value > otherInt.Value ? 1 : 0);
-        }
-
-        public override SValue GreaterThanEqu(SValue other) {
-            if (other is not SInt otherInt) throw new Exception("Can not perform GreaterThanEquCheck on SInt and " + other.GetType().Name);
-            return new SInt(Value >= otherInt.Value ? 1 : 0);
-        }
-
-        public override SValue CastToBuiltin(SBuiltinType other) {
-            switch(other) {
-                case SBuiltinType.Int:
-                    return new SInt(Value);
-                case SBuiltinType.Float:
-                    return new SFloat(Value);
-                default: throw CastInvalid("native " + other.ToString());
-            }
-        }
-
-        public override bool IsTruthy() {
-            return Value == 1;
-        }
-
-        public override string ToString() {
-            return $"<{GetType().Name} value={Value}>";
-        }
-
-        public override SString ToSpagString() {
-            return new SString(Value.ToString());
-        }
-    }
-
-    public class SFloat : SValue {
-        public override SBuiltinType BuiltinName => SBuiltinType.Float;
-        public float Value { get; set; }
-
-        public SFloat() {
-            Value = 0;
-        }
-
-        public SFloat(float value) {
-            Debug.WriteLine("Setting to " + value);
-            Value = value;
-            Debug.WriteLine("Set to " + Value);
-        }
-
-        public override SValue Add(SValue other) {
-            if (other is not SFloat otherInt) throw new Exception("Can not perform Add on SFloat and " + other.GetType().Name);
-            return new SFloat(Value + otherInt.Value);
-        }
-
-        public override SValue Sub(SValue other) {
-            if (other is not SFloat otherInt) throw new Exception("Can not perform Sub on SFloat and " + other.GetType().Name);
-            return new SFloat(Value - otherInt.Value);
-        }
-
-        public override SValue Mul(SValue other) {
-            if (other is not SFloat otherInt) throw new Exception("Can not perform Mul on SFloat and " + other.GetType().Name);
-            return new SFloat(Value * otherInt.Value);
-        }
-
-        public override SValue Div(SValue other) {
-            if (other is not SFloat otherInt) throw new Exception("Can not perform Div on SFloat and " + other.GetType().Name);
-            return new SFloat(Value / otherInt.Value);
-        }
-
-        public override SValue Mod(SValue other) {
-            if (other is not SFloat otherInt) throw new Exception("Can not perform Mod on SFloat and " + other.GetType().Name);
-            return new SFloat(Value % otherInt.Value);
-        }
-
-        public override SValue Equals(SValue other) {
-            if (other is not SFloat otherInt) throw new Exception("Can not perform EqualsCheck on SFloat and " + other.GetType().Name);
-            return new SInt(Value == otherInt.Value ? 1 : 0);
-        }
-
-        public override SValue CastToBuiltin(SBuiltinType other) {
-            switch (other) {
-                case SBuiltinType.Int:
-                    return new SInt((int)Value);
-                case SBuiltinType.Float:
-                    return new SFloat(Value);
-                default: throw CastInvalid("native " + other.ToString());
-            }
-        }
-
-        public override bool IsTruthy() {
-            return Value == 1;
-        }
-
-        public override string ToString() {
-            return $"<{GetType().Name} value={Value}>";
-        }
-
-        public override SString ToSpagString() {
-            return new SString(Value.ToString());
-        }
-    }
-
-    public class SString : SValue
-    {
-        public string Value { get; set; }
-        public override SBuiltinType BuiltinName => SBuiltinType.String;
-
-        public SString() { }
-        public SString(string value)
-        {
-            Value = value;
-        }
-
-        public override SString ToSpagString()
-        {
-            return new SString(Value);
-        }
-
-        public override string ToString()
-        {
-            return $"<SString Value={Value}>";
-        }
-
-        public override SValue Add(SValue other)
-        {
-            if (other is not SString @string) throw new NotImplementedException();
-            return new SString(Value + @string.Value);
-        }
-
-        public override bool IsTruthy() {
-            return Value != null;
-        }
-    }
-
-    public class SList : SValue {
-        public List<SValue> Value { get; set; } = new();
-        public override SBuiltinType BuiltinName => SBuiltinType.List;
-
-        public SList() { }
-
-        public override SString ToSpagString() {
-            return new SString("[" + string.Join(", ", Value.Select((v) => v.ToSpagString().Value)) + "]");
-        }
-
-        public override string ToString() {
-            return $"<SString Value={string.Join(", ", Value)}>";
-        }
-
-        public override SValue Idx(SValue other) {
-            if (other is not SInt otherInt) throw new Exception("Can only index SList with integers, got " + other.GetType().Name);
-
-            if (otherInt.Value < 0 || otherInt.Value > Value.Count - 1) throw new Exception("Out of bounds access. SList had " + Value.Count + " elements, but index " + otherInt.Value + " was accessed");
-            return Value[otherInt.Value];
-        }
-
-        public override bool IsTruthy() {
-            return Value != null;
-        }
-    }
-
-    public struct PrimitiveCompatibility
-    {
-        public SBuiltinType LHS, RHS;
-        public SyntaxType Operator;
-
-        public PrimitiveCompatibility(SBuiltinType lhs, SyntaxType op, SBuiltinType rhs)
-        {
-            LHS = lhs;
-            Operator = op;
-            RHS = rhs;
-        }
-    }
-
-    public enum NodeType
-    {
-        Return,
-        BinaryExpression,
-        Token,
-        BooleanLiteral,
-        Block,
-        Continue,
-        Break,
-        InitVariable,
-        AssignVariable,
-        UnaryExpression,
-        Dot,
-        Call,
-        IntLiteral,
-        FloatLiteral,
-        StringLiteral,
-        Identifier,
-        List,
-        If,
-        For,
-        Cast,
-        While,
-        FunctionDefinition,
-        NativeImport
-    }
-
-    public enum SBuiltinType
-    {
-        String,
-        Int,
-        Float,
-        List,
-        Null,
-        NativeFunc,
-        Function,
-        NativeLibraryImporter,
-    }
 }
